Add files via upload
I've had this idea for a little while now so I thought giving it a test and using it as practice would be good. I'm pretty happy with how it came out but I also see a few spots where it can be done better. See the comments in the code if you care.